Get.map.size {iNZightMaps}	R Documentation

Get the size of the map in pixels

Description

compute the size and the zoom level that needs for request a new map.

Usage

Get.map.size(latR, lonR, SCALE)

Arguments

latR
a numeric vector of length 2, the range of Latitude

lonR
a numeric vector of length 2, the range of Longitude

SCALE
variable from GetMap, use the API's scale parameter to return higher-resolution map images. The scale value is multiplied with the size to determine the actual output size of the image in pixels, without changing the coverage area of the map

Details

since the size of the map should not greater than 640, the size and the zoom needs to be transform before pass to the getNewMap function.

Value

a list that contain the size of the map(in pixels), and the zoom level

Author(s)

Jason Wen

Examples

r.center = runif(4,min = -90,max = 90)
SCALE = 2
Get.map.size(r.center[1:2],r.center[3:4],SCALE)

[Package iNZightMaps version 2.0-0 Index]