drawing.features {iNZightMaps}	R Documentation

drawing features

Description

a function for drawing bar char, display the value and/or region name

Usage

drawing.features(bar.obj, latlon, cols, shade.each, region.name, data.region,
  value, name, center.x, center.y, y.shift = 0.5)

Arguments

bar.obj
a bar object see bar.coor

latlon
a n*2 numeric matrix the first column specifies the latitudes and the second column specifies the longitudes

cols
a color character strings vector

shade.each
a numeric vector

region.name
a character vector

value
a numeric vector

name
a character vector

center.x
a numeric value

center.y
a numeric value

y.shift
a numeric value

Author(s)

Jason

[Package iNZightMaps version 2.0-0 Index]