findBestMatch {iNZightMaps}	R Documentation

Find the pair of variables from the map and data that have the highest number of matches. This function returns in form: (data variable, map variable)

Description

Find the pair of variables from the map and data that have the highest number of matches. This function returns in form: (data variable, map variable)

Usage

findBestMatch(data, map.data)

[Package iNZightMaps version 2.0-0 Index]