getNewMap {iNZightMaps}	R Documentation

Get an new map object

Description

Get a new map and save it into Global Env

Usage

getNewMap(lat.lim, lon.lim, SCALE, type = c("roadmap", "mobile", "satellite",
  "terrain", "hybrid", "mapmaker-roadmap", "mapmaker-hybrid"), zoom)

Arguments

lat.lim
the range of longitude

lon.lim
the range of latitude

SCALE
variable from GetMap, use the API's scale parameter to return higher-resolution map images. The scale value is multiplied with the size to determine the actual output size of the image in pixels, without changing the coverage area of the map

zoom
variable from GetMap, Google maps zoom level.

variable
from GetMap, type defines the type of map to construct. There are several possible maptype values, including satellite, terrain, hybrid, and mobile.

Details

a wrap function of GetMap from package 'RgoogleMaps'

Value

a map object that assign as 'golbal.object'

Author(s)

Jason Wen

[Package iNZightMaps version 2.0-0 Index]