iNZightMap {iNZightMaps}	R Documentation

Create an iNZight Map Object

Description

Create an iNZight Map Object and pass to the plot

Usage

iNZightMap(lat, lon, data, name = deparse(substitute(data)))

## S3 method for class 'inzightmap'
plot(x, opacity, type = c("roadmap", "mobile",
  "satellite", "terrain", "hybrid", "mapmaker-roadmap", "mapmaker-hybrid"), ...,
  env = parent.frame())

Arguments

lat
a formula specifying the name of the latitude variable in data

lon
a formula specifying the name of the longitude variable in data

data
a data frame

name
the name of the data set

x
an inzightmap object

opacity
character or expression of the variable name to code point opacity by

type
the type of map to download from Google

...
additional arguments passed to iNZightPlot

Details

The plot will download a map that contains all of the points. The size is limited by the fixed zoom levels provided by Google's Static Maps API.

Value

data frame object with class inzightmap

Methods (by generic)

plot: Plot an inzightmap object

Author(s)

Tom Elliott

[Package iNZightMaps version 2.0-0 Index]