iNZightMapPlot {iNZightMaps}	R Documentation

Create iNZightMapPlot object

Usage

iNZightMapPlot(data, map, type, ...)

iNZightMapPlotRegion(data, map, by.data, by.map, simplification.level = 0.01,
  multiple.obs = FALSE, sequence.var = NULL, agg.type = "last")

Arguments

data
Dataset with values for rows of the map object

map
An sf object containing a row for each feature of the map

type
What type of map is being passed in. Only "region" is implemented right now.

...
Extra arguments required for the type of map being created.

by.data
Variable name in the dataset that will be matched to by.map in the map

by.map
Variable name in the map that will be matched to by.data in the dataset

simplification.level
How much should the map be simplified by? (see st_simplify)

multiple.obs
Does the dataset have multiple observations for each region of the map (i.e. observations from multiple years)

sequence.var
If multiple.obs = TRUE, which variable identifies the different observations for a region?

agg.type
If multiple.obs = TRUE, which aggregation should be used to produce one observation for each region.

data
Dataset with values for rows of the map object

map
An sf object containing a row for each feature of the map

Functions

iNZightMapPlotRegion: Constructs a iNZightMapPlot using region values.

[Package iNZightMaps version 2.0-0 Index]