iNZightShapeMap {iNZightMaps}	R Documentation

Create an iNZight Shape Map Object

Description

Create an iNZightShapeMap object

Usage

iNZightShapeMap(location, shp.region, data.region, data)

Arguments

location
the shape file

shp.region
a character value, the column name in the region/country column of the shp file

data.region
a character value, the column name in the region/country column of the data set

data
the data set

the
variable that display in the map

Details

details ....

Value

an iNZight Shape Map Object

Author(s)

Tom Elliott

[Package iNZightMaps version 2.0-0 Index]