innerLim {iNZightMaps}	R Documentation

inner limit

Description

Calculate the inner limit.

Usage

innerLim(obj, d.region)

Arguments

obj
an inzshapemap object.

d.region
a character vector that specify the region or country.

Details

the limit is computed by calculate the limit of the given region.

Value

a numeric vector with length 4, the inner limit/bbox of the map.

Author(s)

Jason Wen

[Package iNZightMaps version 2.0-0 Index]