is.google.map {iNZightMaps}	R Documentation

Is Google Map?

Description

The coordinates should be within on the same range of Google Map's. i.e. longitude = [-180,180], latitude = [-90,90]

Usage

is.google.map(lat, lon)

Arguments

lat
a numeric vector

lon
a numeric vector

Value

return an logical value that tells wheater the coordinates are all within the same range of google map's or not.

Author(s)

Jason Wen

[Package iNZightMaps version 2.0-0 Index]