latlon.xy {iNZightMaps}	R Documentation

latlon.xy

Description

Transform the latitude and longitude into xy

Usage

latlon.xy(latlon, map)

Arguments

latlon
the data set that use for plotting, the first cloumn needs to be latitude and the second column needs to be longitude.

map
a map object from RgoogleMaps package.

Details

wrap function for LatLon2XY from 'RgoogleMaps' package

Value

a list that contain the tramsformed x and y.

Author(s)

Jason Wen

[Package iNZightMaps version 2.0-0 Index]