lim.inside {iNZightMaps}	R Documentation

identify the which set of latitude and longitude are within the limit.

Description

is the latitude and longitude within the limit?

Usage

lim.inside(x, lim)

Arguments

x
a numeric matrix of n*2 dimension.

lim
a numeric vector of length 4.

Value

a logical vector of length of n, indicates which row in x is within in the limit.

Author(s)

Jason Wen

Examples

x = cbind(runif(100,-50,50),runif(100,-100,100))
lim = c(-25,25,-80,80)
lim.inside(x,lim)

[Package iNZightMaps version 2.0-0 Index]