lon.rescale {iNZightMaps}	R Documentation

transform the latitude

Description

transform latitude inorder to get the biggest map as possibile

Usage

lon.rescale(lon)

Arguments

lon
a numeric value or a numeric vector of longitude

Details

transform the longitude

Value

a new longitude value or vector that used for plotting in map.

Author(s)

Jason Wen

[Package iNZightMaps version 2.0-0 Index]