matchVariables {iNZightMaps}	R Documentation

Match two vectors

Description

Match two vectors

Usage

matchVariables(data.vect, map.vect)

Arguments

data.vect
Vector containing the dataset variable

map.vect
Vector containing the map object variable

Value

A list containing: the dataset variable vector with duplicates removed, which values in the dataset and map variable had a match, how many matches occured overall and if there was multiple observations for any region.

[Package iNZightMaps version 2.0-0 Index]