needNewMap {iNZightMaps}	R Documentation

Need New Map

Description

Do we need a new map or not

Usage

needNewMap(bbox, window, size, SCALE, type)

Arguments

bbox
a numeric vector of length 4, the range of the pervious latitude and longitude

window
a numeric vector of length 2, the size of the pervious window

SCALE
a numeric vector of length 1, the scale of the pervious map

type
a character vector of length 1, type the type of the pervious map

sized
a numeric vector of length 2, the size of the pervious map

Details

A function for checking wheather we need to download a new map from google or not, it checks the input with the global inzight objects, if something if not matched, then return TRUE, otherwise FALSE.

Value

Logical value TRUE/FALSE TRUE = something are not matched, FALSE = the pervious map is ok for re-use.

Author(s)

Jason Wen

Examples

needNewMap()

[Package iNZightMaps version 2.0-0 Index]