outerLim {iNZightMaps}	R Documentation

outer limit

Description

Calculate the outer limit.

Usage

outerLim(obj, lim, ignore.region = c("Russia", "Antarctica"))

Arguments

obj
an inzshapemap object.

lim
a numeric vector of length 4.

ignore.region
a character value or vector, specify which regions are been ignored when calculate the limit.

Details

the limit is computed by calculate the limit of the given inner limit. If there is a region been cut by the inner limit, then the outer limit will extend up to the limit of the region.

Value

a numeric vector with length 4, the outer limit/bbox of the map.

Author(s)

Jason Wen

Examples

los = 'C:/Users/yeamin/Documents/GitHub/iNZightMaps/data/world.rds'
lod = 'C:/Users/yeamin/Desktop/Gapminder-2008.csv'
data = read.csv(lod,skip = 1)
obj = iNZightShapeMap(los, data.region = 'Country', data = data)
lim = c(-40,40,-40,40)
outerLim(obj,lim)

[Package iNZightMaps version 2.0-0 Index]