plot.iNZightMapPlot {iNZightMaps}	R Documentation

Plot an iNZightMapPlot

Description

Plot an iNZightMapPlot

Usage

## S3 method for class 'iNZightMapPlot'
plot(obj, colour.var = NULL, size.var = NULL,
  alpha.var = NULL, fill.const = NULL, colour.const = NULL,
  size.const = 1, alpha.const = 1, facet = NULL, multiple.vars = FALSE,
  main = NULL, xlab = "Longitude", ylab = "Latitude",
  axis.labels = TRUE, datum.lines = TRUE, darkTheme = NULL,
  projection = "Default", palette = NULL, label.title = "",
  aggregate = FALSE, current.seq = NULL, sparkline.type = "Absolute",
  scale.limits = NULL, ci.plot = FALSE, regions.to.plot = NULL,
  keep.other.regions = TRUE, label.var = NULL, scale.label = 1,
  scale.axis = 1)

Arguments

obj
iNZightMapPlot object

colour.var
Variable to colour the regions or points by

size.var
If plotting a map of points, a variable to scale the points by

size.const
Size of plotted points (ignored if plotting regions or size.var is also passed)

alpha.const
Alpha value of the underlying region map when plotting points

multiple.vars
Are multiple variables being plotted?

main
Main title for the plot

xlab
x-axis label

ylab
y-axis label

axis.labels
Should the x- and y-axis labels be plotted

datum.lines
Should the datum lines be plotte

darkTheme
Enable dark background

projection
Either "Default" or a proj4string

palette
Palette to use

aggregate
Is the plot an aggregate

current.seq
Current value of the sequence variable or aggregation

sparkline.type
Either "Absolute" or "Relative"

scale.limits
Limits for the legend scale

[Package iNZightMaps version 2.0-0 Index]