plot.inzightshapemap {iNZightMaps}	R Documentation

Create an iNZight Shape Map Object

Description

Create an iNZightShapeMap object

Usage

## S3 method for class 'inzightshapemap'
plot(x, variable, col.fun = "hue",
  transform = "linear", col.offset = 0.2, col = "red",
  na.fill = "#F4A460", full.map = TRUE, extend.ratio = 1, name = FALSE,
  zoom = 1, zoom.center = c(NA, NA), ...)

Arguments

x
the iNZight Shape Map Object

col.fun
a character value

full.map
logical value.

variablie
the variable or the column name in the data

region
a variable or the column name of the region column

data
the data set

Details

details ....

Author(s)

Tom Elliott

[Package iNZightMaps version 2.0-0 Index]