win.ratio {iNZightMaps}	R Documentation

win.ratio

Description

compute and return the xlim and ylim within aspect ratio

Usage

win.ratio(xlim, ylim)

Arguments

xlim
a numeric vector of length 2

ylim
a numeric vector of length 2

Value

a numeric vector of length 4, the first two componets specified the new xlim and the last two componets specified the new ylim.

Author(s)

Jason

[Package iNZightMaps version 2.0-0 Index]